[blockchain] refs #114 - Run two skycoin nodes for tests, one for live transactions tests and another for the rest of the tests.
