+++ tests/__init__.py
@@ -2,19 +2,11 @@
 """Default test suite for the Skycoin Lykke integration.
 """
 
-import docker, docker.errors
 import logging
 import os.path
-import time
 import uuid
-import socket, errno
-from tempfile import gettempdir
+from .service_manager import *
 
-service_prefix = 'lykke_sky'
-services_started = dict()
-
-logging.basicConfig()
-log = logging.getLogger(service_prefix)
 
 TestSetupError = RuntimeError
 
@@ -25,6 +17,16 @@ def setup():
         skycoin_data_path = os.path.join(
                 os.path.dirname(os.path.realpath(__file__)),
                 *('data/skycoin'.split('/')))
+        testservice_name = testrun_service_name()
+        testsuite_id = str(uuid.uuid1()).replace('-', '')
+        log.info('Starting test run %s' % (testservice_name,))
+        services_started[testservice_name] = (testsuite_id, False)
+        log.info('Initializing service : mongo')
+        init_service('mongo',           '3.6.4-jessie', 27017)
+
+        log.info('Initializing service : redis')
+        init_service('redis',           '3.0.7-alpine', 6379)
+        log.info('Initializing service : skycoin')
         skycoin_params = " -enable-wallet-api=true" \
                          " -db-path=/data/test/blockchain-180.db" \
                          " -download-peerlist=false" \
@@ -32,17 +34,16 @@ def setup():
                          " -db-read-only=true" \
                          " -disable-networking=true" \
                          " -web-interface-port=6420"
-
-        testservice_name = testrun_service_name()
-        testsuite_id = str(uuid.uuid1()).replace('-', '')
-        log.info('Starting test run %s' % (testservice_name,))
-        services_started[testservice_name] = (testsuite_id, False)
-        log.info('Initializing service : mongo')
-        init_service('mongo',           '3.6.4-jessie', 27017)
-        log.info('Initializing service : redis')
-        init_service('redis',           '3.0.7-alpine', 6379)
-        log.info('Initializing service : skycoin')
         init_service('skycoin/skycoin', 'develop',      6420, command=skycoin_params,
+            volumes={skycoin_data_path: {'bind':'/data/test', 'mode' : 'rw'}})
+        skycoin_params = " -enable-wallet-api=true" \
+                         " -db-path=/data/test/blockchain-180.db" \
+                         " -download-peerlist=false" \
+                         " -rpc-interface=true" \
+                         " -db-read-only=true" \
+                         " -disable-networking=true" \
+                         " -web-interface-port=6421"
+        init_service('skycoin/skycoin', 'develop',      6421, command=skycoin_params,
             volumes={skycoin_data_path: {'bind':'/data/test', 'mode' : 'rw'}})
     except:
         log.error('Error found in test suite setup')
@@ -60,112 +61,3 @@ def teardown():
         if service_id.startswith('_'.join((service_prefix, testrun_id()))) and launched:
             docker_dispose(service_id)
     del services_started[testservice_name]
-
-def testrun_service_name():
-    """Name to register this test run as a service.
-
-    Depends on operating-system process ID.
-    """
-    return '%s_%d' % (service_prefix, os.getpid())
-
-def testrun_id():
-    """Global identifier of this test run.
-    """
-    return services_started[testrun_service_name()][0]
-
-def init_service(service_name, version, default_port, command=None, volumes = None):
-    """Ensure service is running. If not available run it with Docker.
-
-    service_name : The name of the official Docker image used to run
-                   the target service.
-    default_port : Default service port. Usually matches the value
-                   EXPOSE'd in Dockerfile.
-    """
-    try:
-        port = int(os.getenv('PORT_' + service_name.upper(), default_port))
-    except ValueError:
-        port = default_port
-
-    launched = False
-    service_id = docker_service_name(service_name)
-    if not is_listening_at(port):
-        docker_run(service_name, tag=version,
-                name=service_id ,
-                ports={'%s/tcp' % (default_port,): port},
-                command=command,
-                volumes=volumes)
-        launched = True
-    services_started[service_id] = (port, launched)
-
-def docker_service_name(service_name):
-    """Name of the container running a service used by this test run, if any.
-    """
-    return '_'.join((service_prefix, testrun_id(), service_name.replace('/', '_')))
-
-def docker_run(service_name, name, **docker_options):
-    """Run service in a Docker container.
-    """
-    client = docker.from_env()
-
-    image_tag = docker_options.get('tag', 'latest')
-    volumes = docker_options.get('volumes', {})
-#    try:
-#        image = client.images.get(service_name)
-#        log.info('Image %s:%s available locally' % (service_name, image_tag))
-#    except:
-#        log.warning('Image %s:%s not found locally ... pulling' % (service_name, image_tag))
-#        client.images.pull(service_name, image_tag)
-    client.containers.run('%s:%s' % (service_name, image_tag), detach=True, name=name,
-            ports=docker_options.get('ports', dict()), command=docker_options.get('command'),
-            volumes=volumes)
-    time.sleep(1.0)
-
-def docker_dispose(service_id):
-    """Remove container used to run service
-    """
-    client = docker.from_env()
-    log.debug('Stopping Docker container ... ' + service_id)
-    container = client.containers.get(service_id)
-    container.stop()
-    container.wait()
-    container.remove()
-
-def wait_for_all(services_started, max_retries=5):
-    """Wait for all services to be available.
-    """
-    log.info('Waiting for services started: %s',
-            ', '.join(key
-                for key, (_, launched) in services_started.items()
-                if launched))
-    step = max_retries
-    pending = services_started.keys()
-    timeout = 3.0
-    time.sleep(timeout)
-    # FIXME: Signal instead of busy wait
-    while step > 0 and len(pending) > 0:
-        time.sleep(timeout)
-        pending = [(service_name, port)
-                for service_name, (port, launched) in services_started.items()
-                if launched and not is_listening_at(port)]
-        timeout = 3
-        step -= 1
-    if len(pending) > 0:
-        log.warning('Some services could not be started: %s',
-                ', '.join('%s at %s' % entry for entry in pending))
-        return False
-    log.info('All services started')
-    return True
-
-def is_listening_at(port):
-    """Check whether there's a service listening at given port.
-    """
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    try:
-        s.bind(("127.0.0.1", port))
-    except socket.error as e:
-        if e.errno == errno.EADDRINUSE:
-            log.info("Listening ")
-        else:
-            # something else raised the socket.error exception
-            print(e)
-    s.close()
